+++ src/client/sync/mod.rs
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright (c) 2017-2023 slowtec GmbH <post@slowtec.de>
+// SPDX-FileCopyrightText: Copyright (c) 2017-2024 slowtec GmbH <post@slowtec.de>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 //! Synchronous Modbus client
@@ -11,7 +11,7 @@ pub mod tcp;
 
 use std::{future::Future, io::Result, time::Duration};
 
-use futures::future::Either;
+use futures_util::future::Either;
 
 use crate::{frame::*, slave::*};