+++ src/codec/tcp.rs
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright (c) 2017-2023 slowtec GmbH <post@slowtec.de>
+// SPDX-FileCopyrightText: Copyright (c) 2017-2024 slowtec GmbH <post@slowtec.de>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use std::io::{Error, ErrorKind, Result};
@@ -295,7 +295,7 @@ mod tests {
             assert_eq!(buf[5], 0x6);
             assert_eq!(buf[6], UNIT_ID);
 
-            _ = buf.split_to(7);
+            drop(buf.split_to(7));
             let pdu: Bytes = req.try_into().unwrap();
             assert_eq!(buf, pdu);
         }